chore: lighthouse 설정을 desktop과 mobile로 분리 (SP-668)
